Startup Speedboost
Changed the program so that it checks if the cave.html file has been cahched. If it finds the cached map file, then it skips creating a new map file. This speeds up the start of the program significantly. The program is still slow when displaying all of the caves at once.